requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Authlib==0.15.4
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-salesforce-oauth==0.3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
fcm-django==0.3.10
google-api-core==1.30.0
google-auth==1.31.0
google-cloud==0.34.0
google-cloud-vision==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.0
gunicorn==20.1.0
heroku==0.1.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
msgpack==1.0.2
numpy==1.21.0
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==20.9
pandas==1.3.0
patsy==0.5.1
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
pyfunctional==1.4.3
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
secure-smtplib==0.1.1
service-identity==21.1.0
simple-salesforce==1.11.2
simplejson==3.17.2
six==1.16.0
sklearn==0.0
sklearn-utils==0.0.15
sqlparse==0.4.1
statsmodels==0.12.2
tabulate==0.8.9
threadpoolctl==2.2.0
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.5
websocket-client==1.1.0
whitenoise==5.2.0
zope.interface==5.4.0
